requirements.txt
astroid==2.4.2
colorama==0.4.4
Django==2.2.17
django-birthday==0.1.3
isort==5.6.4
lazy-object-proxy==1.5.2
mccabe==0.6.1
memory-profiler==0.58.0
mysql-connector-python==8.0.22
numpy==1.19.4
Pillow==8.0.1
protobuf==3.14.0
psutil==5.8.0
pylint==2.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
wrapt==1.12.1
